dll_0B_func09: signed (int) cast on randomGetRange for dscale division (96.44->96.52)
modgfx.h declares randomGetRange as u32, so dscale / randomGetRange(1,10) emitted
the unsigned int->float conversion (stw r3; lis 17200; lfd). Target uses the signed
conversion (xoris r0,r3,32768; ...). Cast the call result to (int) at this site to
force the signed bias path, matching target. #10 signed-vs-unsigned conversion.

(cherry picked from commit 50c79c2a4527cae0704f88613806250be77be737)

+++ src/main/dll/dll_000B_dll0b.c
@@ -2344,7 +2344,7 @@ int dll_0B_func09(void* a0, int a1, int a2, u8 a3, void* a4)
         if (*(int*)((char*)p[slot] + 0xa4) & 0x400000)
         {
             dscale = lbl_803DF458 * *(f32*)((char*)p[slot] + 0xd4);
-            xf.scale = dscale + dscale / randomGetRange(1, 10);
+            xf.scale = dscale + dscale / (int)randomGetRange(1, 10);
         }
         else
         {
